Remove .idata and move .rsrc forward: size reduced to < 1.2 MiB
All traces of SecuROM are now completely gone, and the chasm of zeros
where its sections used to live has been removed. The vestigial .idata
section was removed, as it wasn't actually being used; and the .rsrc
section was moved from its former location at RA 2F2000 / VA F10000 to
its new location at RA 125000 / VA D43000.

The PE header data directory offsets were updated; the offsets in the
import table for openloco.dll were adjusted; the offsets in the resource
table were adjusted; and the jump at the beginning of WinMain to the
StartOpenLoco import was adjusted.

This has been verified and tested, and appears to work entirely
correctly.