Implement strlcpy and strlcat internally
Hypothetically we should be able to just get these from the standard
library or somewhere easy like that; but glibc is belligerent and
refuses to ever add the functions, and MSVCRT doesn't care because they
have their own (not functionally identical) collection of `_s`
functions.

In theory we could link with libbsd on Linux, and set up the right
preprocessor defs on macOS before including standard library headers,
and then... do something for Windows; but frankly I don't want to deal
with all of that garbage.

The continued lack of standardization of strlcpy and strlcat is
disturbing.