Checks Scale Factor is not outside expected range
Implemented a check when the Plus or Minus button is pressed on the Options window to change the Scale Factor to make sure that the new scale factor is actually different to the current one.
If the new scale factor has not changed (ie. it's outside the expected range) we cancel the function instead of refreshing the Screen/UI and causing a purple flash.