Fix #409. Incorrect refund cost when deleting signals.
The cost had incorrectly been marked as unsigned which caused a comparison to fail. It looks like there is an original bug that caused signal deletion to return a pointer to the stack instead of the cost which due to the incorrect cost check caused a very large number to be taken as the cost.