Fix #429; simplify SDL window mode change functions (#431)
* Function for easily changing display type and size.

* Simplify the getDisplayResolutionByMode code.

* Simplify setDisplayMode; only change config if successful.

* Mark legacy config resolutions as signed.

The legacy config resolutions are initialised to -1. Unsigned, this is interpreted as 65535. By marking them as signed, we can continue to use the > 0 sanity check as implemented.

* Change default window size to 800x600.

* Use SetWindowSize to prevent resolution change failures

I don't know fully why this has to be done this way but this follows the way OpenRCT2 does it and prevents a resolution change failure on my setup

* Fix windows display issue by using hack similar to OpenRCT2

Due to the way that Windows uses its own scaling for high dpi monitors we require to use this hack to set the resolution correctly.

Co-authored-by: Aaron van Geffen <aaron@aaronweb.net>
Co-authored-by: duncanspumpkin <duncans_pumpkin@hotmail.co.uk>