Fix #491: Station/city name labels are hidden from viewport when saving (#551)
Mistake made in viewport create function meant that an un-implemented function would not get a pointer to the newly created viewport but a pointer to the main windows viewport. When it set the flags for the new viewport it instead set the flags for the main viewport. Fixed by passing back the viewport to the correct register.