Fix stepper decrement button width.
When introducing the stepper widgets to OpenLoco, I decided to give both buttons an equal width
of 12px. The minus glyph on the decrement button is not optically aligned in this configuration,
however. To remedy this, this commit increases the width of just the decrement button to 13px.
As a result, the glyph is now aligned properly.