Fix #660. Missing += rowHeight caused drawing to perform incorrectly
This bit of code is meant to reduce the amount of drawing taking place. Without the += rowHeight the code would do its skipping of drawing but then when it was meant to draw it would be at the wrong y coordinate. This would cause a flickering when items behind the window caused redraws of parts of the screen.