Fix load save window crashing Windows builds (#888)
Sub_441FA7 in a lot of situations will cause a crash this prevents just one of them which is gauranteed to happen. It is quite likely that there are other crashes here that will still happen. Full implementation of all callers and the function will need to be done before we can be confident that there is no sources of crashes here