Fix crash on save game open (#1121)
My previous PR, #1118, removed the hooks to the prompt browse window. This was working fine for all except saving games! 
Turns out the toolbar menu was still calling the vanilla function, thereby relying on the hook.

Replacing this call revealed a bug in hiding with the save prompt being initialised to the wrong save path.
This has been addressed as well.