Hook malloc and free a bit higher up in the call stack. Fixes running on Apple M1.
On Apple M1, we can run OpenLoco in Wine32on64 within Rosetta 2.  However, it
seems that for some reason, these malloc and free hooks don't seem to work.
The realloc hook does work, therefore our implementation of realloc() receives
a block of memory that was never allocated by our implementation of malloc();
this causes realloc() to return a nullptr which causes a segfault.

By hooking these malloc and free wrappers, which simply call the malloc and free
implementations we were already hooking, our implementation of malloc and free
is properly called and OpenLoco runs on M1.