Fix #1676, #1702. Scenario list crash/hang on new scenarios
Mistake made in implementation of the editorcontroller meant that invalid currency objects were set as the currency of a scenario meta details. This would cause the game to unload the currency object which would change the layout of image data set for water object which would cause invalid offsets to be set in the palette table which would cause the image draw functions to behave oddly causing crashs or infinite loops.