Fix hang when crashing uncleanly (#1883)
* Fix hang when crashing uncleanly

If we crash uncleanly i.e. not through an exception then the audio will get stuck waiting on a lock that is never free'd. This address's the issue by removing the destructor. If we exit uncleanly the audio will not be shutdown correctly but at least we wont hang. Also added an additional catch all exception handler as its possible we might have thrown an exception that isn't derived from std::exception

* Update changelog