Fix #2377: Error message appearing below window (#2403)
This happens when cloning vehicle with insufficient funds.
What happened was that the result of the clone command is not checked,
a window is opened to that cloned vehicle which actually doesn't
exist anymore at that point. Thanks to some defensive coding that
window actually never appears, but it is enough to prevent the
error message rendering at the correct Z-index.