Fix internal progress bars not showing
The internal progress bars weren't displaying at all due to recent changes in the graphics stack (#1766).
The progress bars use Gfx::renderAndUpdate to refresh the screen, but after this refactor, only
SoftwareDrawingEngine::render was being called, but the SoftwareDrawingEngine::present it introduced was not.
This left the screen not updated.