Merge Sound and Music tabs of Options window
Creating "Sound" and "Jukebox" groups in the tab, and defining new string IDs for text that was altered/replaced/added.

Layout of widgets is the same as in the original tabs, except the music volume label was shifted to the left.

namespace is named AudioTab to avoid name collision with the Audio namespace from the Audio module.