Add Turkish translation (#3515)
All lines listed in the language pack have been translated into Turkish. A small portion of the game that has not yet been modified may be translated in later versions. Due to the coding system and the structure of the Turkish language, there is a minor wording issue in sentences related to “which products factories accept and which they output,” but it is not incomprehensible.